[[ Engine ]] Tweak OS X cursors to work with scroll views
This patch removes a check on OS X which was preventing the cursor from
chnaging when over a native layer. This was having the effect of
stoping the cursor from updating when a native scroll view was used.

Additionally, support has been added to OS X for the platform hand
cursor via id 10.